Minor: move `Column` related tests
